Add test for HomeserverConfiguration build.
