+++ RiotTests/VectorWellKnownTests.swift
@@ -101,39 +101,4 @@ class VectorWellKnownTests: XCTestCase {
             XCTFail("Fail with error: \(error)")
         }
     }
-    
-    func testMXWellKnown() {
-        
-        let expectedJitsiServer = "your.jitsi.example.org"
-        let expectedDeprecatedJitsiServer = "your.deprecated.jitsi.example.org"
-        let expectedE2EEEByDefaultEnabled = true
-        let expectedDeprecatedE2EEEByDefaultEnabled = false
-        
-        let wellKnownDictionary: [String: Any] = [
-            "m.homeserver": [
-                 "base_url": "https://your.homeserver.org"
-            ],
-             "m.identity_server": [
-                 "base_url": "https://your.identity-server.org"
-            ],
-            "im.vector.riot.e2ee" : [
-                "default" : expectedDeprecatedE2EEEByDefaultEnabled
-            ],
-            "im.vector.riot.jitsi" : [
-                "preferredDomain" : expectedDeprecatedJitsiServer
-            ],
-            "io.element.e2ee" : [
-                "default" : expectedE2EEEByDefaultEnabled
-            ],
-            "io.element.jitsi" : [
-                "preferredDomain" : expectedJitsiServer
-            ]
-        ]
-                
-        
-        let wellKnown = MXWellKnown(fromJSON: wellKnownDictionary)
-        
-        XCTAssertEqual(wellKnown?.vc_jitsiPreferredDomain(), expectedJitsiServer)
-        XCTAssertEqual(wellKnown?.vc_isE2EEByDefaultEnabled(), expectedE2EEEByDefaultEnabled)
-    }
 }
